Bug 2030411: move enterprise mock console also for mochitest
